requirements.txt
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.26.0
tiktoken
einops
transformers_stream_generator>=0.0.4
scipy
peft>=0.4.0
huggingface_hub
modelscope
psutil
pynvml
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sseclient-py